settings ctrl cleanup + functions renaming
