+++ src/controllers/settings/settings.ts
@@ -1,52 +1,19 @@
-/* eslint-disable no-underscore-dangle */
-import EmittableError from '../../classes/EmittableError'
-import { AMBIRE_ACCOUNT_FACTORY } from '../../consts/deploy'
-import { networks } from '../../consts/networks'
 import { Key } from '../../interfaces/keystore'
-import {
-  Network,
-  NetworkFeature,
-  NetworkId,
-  NetworkInfo,
-  NetworkInfoLoading
-} from '../../interfaces/network'
-import {
-  AccountPreferences,
-  CustomNetwork,
-  KeyPreferences,
-  NetworkPreference,
-  NetworkPreferences,
-  RPCProviders
-} from '../../interfaces/settings'
+import { AccountPreferences, KeyPreferences } from '../../interfaces/settings'
 import { Storage } from '../../interfaces/storage'
-import { getFeaturesByNetworkProperties, getNetworkInfo } from '../../libs/settings/settings'
 import { isValidAddress } from '../../services/address'
-import { getRpcProvider } from '../../services/provider'
 import EventEmitter, { Statuses } from '../eventEmitter/eventEmitter'
 
-const STATUS_WRAPPED_METHODS = {
-  addCustomNetwork: 'INITIAL',
-  updateNetworkPreferences: 'INITIAL'
-} as const
+const STATUS_WRAPPED_METHODS = {} as const
 
 export class SettingsController extends EventEmitter {
   accountPreferences: AccountPreferences = {}
 
   keyPreferences: KeyPreferences = []
 
-  providers: RPCProviders = {}
-
-  #networkPreferences: NetworkPreferences = {}
-
   #storage: Storage
 
   statuses: Statuses<keyof typeof STATUS_WRAPPED_METHODS> = STATUS_WRAPPED_METHODS
-
-  networkToAddOrUpdate: {
-    chainId: Network['chainId']
-    rpcUrl: string
-    info?: NetworkInfoLoading<NetworkInfo>
-  } | null = null
 
   constructor(storage: Storage) {
     super()
@@ -55,128 +22,13 @@ export class SettingsController extends EventEmitter {
     this.#load()
   }
 
-  #setProvider(network: Network, newRpcUrls: string[], selectedRpcUrl?: string) {
-    const provider = this.providers[network.id]
-
-    // Only update the RPC if the new RPC is different from the current one
-    // or if there is no RPC for this network yet.
-    // eslint-disable-next-line no-underscore-dangle
-    if (!provider || provider?._getConnection().url !== selectedRpcUrl) {
-      const oldRPC = this.providers[network.id]
-
-      if (oldRPC) {
-        // If an RPC fails once it will try to reconnect every second. If we don't destroy the old RPC
-        // it will keep trying to reconnect forever.
-        oldRPC.destroy()
-      }
-
-      this.providers[network.id] = getRpcProvider(newRpcUrls, network.chainId, selectedRpcUrl)
-    }
-  }
-
-  get networks(): (Network & (NetworkPreference | CustomNetwork))[] {
-    // set the custom networks that do not exist in ambire-common networks
-    const customPrefIds = Object.keys(this.#networkPreferences)
-    const predefinedNetworkIds = networks.map((net) => net.id)
-    const customNetworkIds = customPrefIds.filter((x) => !predefinedNetworkIds.includes(x))
-    const customNetworks = customNetworkIds.map((id: NetworkId) => {
-      // @ts-ignore
-      // checked with the logic for customNetworkIds
-      const customNetwork: CustomNetwork = this.#networkPreferences[id]
-      const features: NetworkFeature[] = []
-
-      return {
-        id,
-        unstoppableDomainsChain: 'ERC20',
-        name: customNetwork.name,
-        nativeAssetSymbol: customNetwork.nativeAssetSymbol,
-        rpcUrls: customNetwork.rpcUrls,
-        chainId: customNetwork.chainId,
-        explorerUrl: customNetwork.explorerUrl,
-        erc4337: customNetwork.erc4337 ?? { enabled: false, hasPaymaster: false },
-        isSAEnabled: customNetwork.isSAEnabled ?? false,
-        areContractsDeployed: customNetwork.areContractsDeployed ?? false,
-        isOptimistic: customNetwork.isOptimistic ?? false,
-        rpcNoStateOverride: customNetwork.rpcNoStateOverride ?? true,
-        hasDebugTraceCall: customNetwork.hasDebugTraceCall ?? false,
-        platformId: customNetwork.platformId ?? '',
-        nativeAssetId: customNetwork.nativeAssetId ?? '',
-        flagged: customNetwork.flagged ?? false,
-        feeOptions: customNetwork.feeOptions ?? {
-          is1559: false
-        },
-        features,
-        hasRelayer: false,
-        hasSingleton: customNetwork.hasSingleton ?? false
-      }
-    })
-
-    const allNetworks = [...networks, ...customNetworks]
-
-    // configure the main networks
-    return allNetworks.map((network) => {
-      const networkPreferences = this.#networkPreferences[network.id]
-
-      // erc4337 settings should not be inherited from networkPreferences
-      // for predefined networks
-      if (
-        predefinedNetworkIds.includes(network.id) &&
-        networkPreferences &&
-        'erc4337' in networkPreferences
-      )
-        delete networkPreferences.erc4337
-
-      const selectedRpcUrl =
-        networkPreferences?.selectedRpcUrl || networkPreferences?.rpcUrls?.[0] || network.rpcUrls[0]
-      this.#setProvider(network, networkPreferences?.rpcUrls || network.rpcUrls, selectedRpcUrl)
-      const finalNetwork = networkPreferences
-        ? {
-            ...network,
-            ...networkPreferences,
-            selectedRpcUrl
-          }
-        : {
-            ...network,
-            selectedRpcUrl
-          }
-
-      const info: NetworkInfo = {
-        isSAEnabled: finalNetwork.isSAEnabled,
-        isOptimistic: finalNetwork.isOptimistic ?? false,
-        rpcNoStateOverride: finalNetwork.rpcNoStateOverride,
-        erc4337: finalNetwork.erc4337,
-        areContractsDeployed: finalNetwork.areContractsDeployed,
-        feeOptions: finalNetwork.feeOptions,
-        hasDebugTraceCall: finalNetwork.hasDebugTraceCall,
-        platformId: finalNetwork.platformId,
-        nativeAssetId: finalNetwork.nativeAssetId,
-        flagged: finalNetwork.flagged ?? false,
-        chainId: finalNetwork.chainId,
-        hasSingleton: finalNetwork.hasSingleton
-      }
-
-      finalNetwork.features = getFeaturesByNetworkProperties(info)
-      return finalNetwork
-    })
-  }
-
-  updateProviderIsWorking(networkId: NetworkId, isWorking: boolean) {
-    this.providers[networkId].isWorking = isWorking
-
-    this.emitUpdate()
-  }
-
   async #load() {
     try {
-      let networkPreferences = {}
-      ;[this.accountPreferences, this.keyPreferences, networkPreferences] = await Promise.all([
+      ;[this.accountPreferences, this.keyPreferences] = await Promise.all([
         // Should get the storage data from all keys here
         this.#storage.get('accountPreferences', {}),
-        this.#storage.get('keyPreferences', []),
-        this.#storage.get('networkPreferences', {})
+        this.#storage.get('keyPreferences', [])
       ])
-
-      this.#networkPreferences = this.#migrateNetworkPreferences(networkPreferences)
 
       this.emitUpdate()
     } catch (e) {
@@ -189,29 +41,11 @@ export class SettingsController extends EventEmitter {
     }
   }
 
-  // eslint-disable-next-line class-methods-use-this
-  #migrateNetworkPreferences(networkPreferencesOldFormat: {
-    [key in NetworkId]: (NetworkPreference | CustomNetwork) & { rpcUrl?: string }
-  }) {
-    const modifiedNetworks: NetworkPreferences = {}
-    // eslint-disable-next-line no-restricted-syntax
-    for (const [networkId, network] of Object.entries(networkPreferencesOldFormat)) {
-      if (network.rpcUrl && !network.rpcUrls) {
-        modifiedNetworks[networkId] = { ...network, rpcUrls: [network.rpcUrl] }
-      } else {
-        modifiedNetworks[networkId] = network
-      }
-    }
-
-    return modifiedNetworks
-  }
-
   async #storePreferences() {
     try {
       await Promise.all([
         this.#storage.set('accountPreferences', this.accountPreferences),
-        this.#storage.set('keyPreferences', this.keyPreferences),
-        this.#storage.set('networkPreferences', this.#networkPreferences)
+        this.#storage.set('keyPreferences', this.keyPreferences)
       ])
     } catch (e) {
       this.emitError({
@@ -342,217 +176,6 @@ export class SettingsController extends EventEmitter {
     this.emitUpdate()
   }
 
-  setNetworkToAddOrUpdate(
-    networkToAddOrUpdate: {
-      chainId: Network['chainId']
-      rpcUrl: string
-    } | null = null
-  ) {
-    if (networkToAddOrUpdate) {
-      this.networkToAddOrUpdate = networkToAddOrUpdate
-      this.emitUpdate()
-
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      getNetworkInfo(networkToAddOrUpdate.rpcUrl, networkToAddOrUpdate.chainId, (info) => {
-        if (this.networkToAddOrUpdate) {
-          this.networkToAddOrUpdate = { ...this.networkToAddOrUpdate, info }
-          this.emitUpdate()
-        }
-      })
-    } else {
-      this.networkToAddOrUpdate = null
-      this.emitUpdate()
-    }
-  }
-
-  // NOTE: used only for adding a new network. We add it to the user's
-  // this.#networkPreferences as one day the custom network can be a part of
-  // Ambire's main networks => the options in it will become a preference
-  async #addCustomNetwork(customNetwork: CustomNetwork) {
-    // mainCtrl.settings.setNetworkToAddOrUpdate
-    if (
-      !this.networkToAddOrUpdate?.info ||
-      Object.values(this.networkToAddOrUpdate.info).some((prop) => prop === 'LOADING')
-    ) {
-      return
-    }
-    const chainIds = this.networks.map((net) => net.chainId)
-    const customNetworkId = customNetwork.name.toLowerCase()
-    const ids = this.networks.map((net) => net.id)
-
-    // make sure the id and chainId of the network are unique
-    if (
-      ids.indexOf(customNetworkId) !== -1 ||
-      chainIds.indexOf(BigInt(customNetwork.chainId)) !== -1
-    ) {
-      throw new EmittableError({
-        message: 'The network you are trying to add has already been added.',
-        level: 'major',
-        error: new Error('settings: addCustomNetwork chain already added (duplicate id/chainId)')
-      })
-    }
-
-    const info = { ...(this.networkToAddOrUpdate.info as NetworkInfo) }
-    const { feeOptions } = info
-
-    // eslint-disable-next-line no-param-reassign
-    delete (info as any).feeOptions
-
-    this.#networkPreferences[customNetworkId] = {
-      ...customNetwork,
-      ...info,
-      ...feeOptions
-    }
-
-    await this.#storePreferences()
-    this.networkToAddOrUpdate = null
-    this.emitUpdate()
-  }
-
-  async addCustomNetwork(customNetwork: CustomNetwork) {
-    await this.withStatus(this.addCustomNetwork.name, () => this.#addCustomNetwork(customNetwork))
-  }
-
-  async removeCustomNetwork(id: NetworkId) {
-    if (networks.find((n) => n.id === id)) return
-
-    delete this.#networkPreferences[id]
-    this.providers?.[id]?.destroy()
-    delete this.providers?.[id]
-    await this.#storePreferences()
-    this.emitUpdate()
-  }
-
-  async #updateNetworkPreferences(
-    networkPreferences: Partial<NetworkPreference>,
-    networkId: NetworkId
-  ) {
-    if (!Object.keys(networkPreferences).length) return
-
-    const networkData = this.networks.find((network) => network.id === networkId)
-
-    const changedNetworkPreferences: NetworkPreference = Object.keys(networkPreferences).reduce(
-      (acc, key) => {
-        if (!networkData) return acc
-
-        // No need to save unchanged network preferences.
-        // Here we filter the network preferences that are the same as the ones in the storage.
-        if (
-          networkPreferences[key as keyof NetworkPreference] ===
-          networkData[key as keyof NetworkPreference]
-        )
-          return acc
-
-        return { ...acc, [key]: networkPreferences[key as keyof NetworkPreference] }
-      },
-      {} as NetworkPreference
-    )
-
-    // Update the network preferences with the incoming new values
-    this.#networkPreferences[networkId] = {
-      ...this.#networkPreferences[networkId],
-      ...changedNetworkPreferences
-    }
-
-    await this.#storePreferences()
-    this.emitUpdate()
-
-    // Do not wait the rpc validation in order to complete the execution of updateNetworkPreferences
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    ;(async () => {
-      // if the rpcUrls have changed, call the RPC and check whether it supports
-      // state overrided. If it doesn't, add a warning
-      if (changedNetworkPreferences.selectedRpcUrl) {
-        if (
-          this.networkToAddOrUpdate?.info &&
-          Object.values(this.networkToAddOrUpdate.info).every((prop) => prop !== 'LOADING')
-        ) {
-          const info = { ...(this.networkToAddOrUpdate.info as NetworkInfo) }
-          const { feeOptions } = info
-
-          // eslint-disable-next-line no-param-reassign
-          delete (info as any).feeOptions
-          this.#networkPreferences[networkId] = {
-            ...this.#networkPreferences[networkId],
-            ...info,
-            ...feeOptions
-          }
-
-          await this.#storePreferences()
-          this.emitUpdate()
-          return
-        }
-
-        // eslint-disable-next-line @typescript-eslint/no-floating-promises
-        getNetworkInfo(
-          changedNetworkPreferences.selectedRpcUrl,
-          this.#networkPreferences[networkId].chainId!,
-          async (info) => {
-            if (Object.values(info).some((prop) => prop === 'LOADING')) {
-              return
-            }
-
-            const { feeOptions } = info as NetworkInfo
-
-            // eslint-disable-next-line no-param-reassign
-            delete (info as any).feeOptions
-            this.#networkPreferences[networkId] = {
-              ...this.#networkPreferences[networkId],
-              ...(info as NetworkInfo),
-              ...feeOptions
-            }
-
-            await this.#storePreferences()
-            this.emitUpdate()
-          }
-        )
-      }
-    })()
-  }
-
-  async updateNetworkPreferences(
-    networkPreferences: Partial<NetworkPreference>,
-    networkId: NetworkId
-  ) {
-    await this.withStatus(this.updateNetworkPreferences.name, () =>
-      this.#updateNetworkPreferences(networkPreferences, networkId)
-    )
-  }
-
-  // NOTE: use this method only for predefined networks
-  async resetNetworkPreference(key: keyof NetworkPreference, networkId: NetworkId) {
-    if (
-      !networkId ||
-      !(networkId in this.#networkPreferences) ||
-      !(key in this.#networkPreferences[networkId])
-    )
-      return
-
-    delete this.#networkPreferences[networkId][key]
-
-    await this.#storePreferences()
-    this.emitUpdate()
-  }
-
-  // call this function after a call to the singleton has been made
-  // it will check if the factory has been deployed and update the
-  // network settings if it has been
-  async setContractsDeployedToTrueIfDeployed(network: Network) {
-    if (network.areContractsDeployed) return
-
-    const provider = this.providers[network.id]
-    const factoryCode = await provider.getCode(AMBIRE_ACCOUNT_FACTORY)
-    if (factoryCode === '0x') return
-
-    this.#updateNetworkPreferences({ areContractsDeployed: true }, network.id).catch(() => {
-      this.emitError({
-        level: 'silent',
-        message: 'Failed to update the network feature for supporting smart accounts',
-        error: new Error(`settings: failed to set areContractsDeployed to true for ${network.id}`)
-      })
-    })
-  }
-
   #throwInvalidAddress(addresses: string[]) {
     return this.emitError({
       message:
@@ -569,9 +192,7 @@ export class SettingsController extends EventEmitter {
   toJSON() {
     return {
       ...this,
-      ...super.toJSON(),
-      networks: this.networks,
-      providers: this.providers
+      ...super.toJSON()
     }
   }
 }
